requirements.txt
pandas
numpy
faiss-cpu
sentence-transformers
rank-bm25
scikit-learn
groq
python-dotenv


